backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.0.16
openai==1.10.0

# Vector Store and Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and Security
python-jose[cryptography]==3.3.0

# Utilities
httpx==0.26.0
tenacity==8.2.3

